cors config update
